remove, refactor, update comments and documentation with human understanding
